split out different action-dropdown pieces a bit
